Adds CSVReader module
